fix(room-http-api): lint修正・E2E実装・テストファイル分割 (Issue #57)
3レビュー指摘を全修正:
1. ruff format + E501 ドキュメント文字列修正でlintクリア
2. tests/e2e/test_room_http_api.py 新規実装
   - TC-E2E-RM-004: 8ステップ Room ライフサイクル一気通貫
   - TC-E2E-RM-005: 不正 UUID パスパラメータ → 422
3. 1056行統合テストを test_room_http_api/ に分割
   (create/list/get/update/archive/agents/validation)
   670行ユニットテストも同様に分割

全1445テスト通過
