--- backend/tests/unit/test_room_http_api.py
@@ -1,670 +1,0 @@
-"""room / http-api ユニットテスト (TC-UT-RM-HTTP-001~006, 010).
-
-Per ``docs/features/room/http-api/test-design.md`` §ユニットテストケース.
-
-Covers:
-  TC-UT-RM-HTTP-001  RoomCreate スキーマ検証 (Q-3)
-  TC-UT-RM-HTTP-002  RoomUpdate スキーマ検証 (Q-3)
-  TC-UT-RM-HTTP-003  AgentAssignRequest スキーマ検証 (Q-3)
-  TC-UT-RM-HTTP-004  RoomResponse / MemberResponse / RoomListResponse シリアライズ (Q-3)
-  TC-UT-RM-HTTP-005  room_invariant_violation_handler [FAIL]/Next: 除去 + kind 分岐 (MSG-RM-HTTP-005/007)
-  TC-UT-RM-HTTP-006  RoomService.__init__ 4 repo 構造 (Q-3)
-  TC-UT-RM-HTTP-010  interfaces/http/ 依存方向静的解析 — ast.walk() 全ノード走査 (Q-3)
-
-Issue: #57
-"""
-
-from __future__ import annotations
-
-import ast
-from datetime import UTC, datetime
-from pathlib import Path
-from typing import Any
-from unittest.mock import MagicMock
-from uuid import uuid4
-
-import pytest
-from pydantic import ValidationError
-
-
-# ---------------------------------------------------------------------------
-# TC-UT-RM-HTTP-001: RoomCreate スキーマ検証
-# ---------------------------------------------------------------------------
-
-
-class TestRoomCreateSchema:
-    """TC-UT-RM-HTTP-001: RoomCreate validates name / description / prompt_kit_prefix_markdown."""
-
-    def test_valid_name_workflow_id_passes(self) -> None:
-        """(a) name="Vモデル開発室", workflow_id=uuid4() → バリデーション通過."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        schema = RoomCreate(name="Vモデル開発室", workflow_id=uuid4())
-        assert schema.name == "Vモデル開発室"
-
-    def test_empty_name_raises(self) -> None:
-        """(b) name="" → min_length=1 違反 → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        with pytest.raises(ValidationError):
-            RoomCreate(name="", workflow_id=uuid4())
-
-    def test_name_too_long_raises(self) -> None:
-        """(c) name='x'*81 → max_length=80 違反 → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        with pytest.raises(ValidationError):
-            RoomCreate(name="x" * 81, workflow_id=uuid4())
-
-    def test_name_80_chars_passes(self) -> None:
-        """Boundary: 80 文字は上限に等しく通過すべき."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        schema = RoomCreate(name="a" * 80, workflow_id=uuid4())
-        assert len(schema.name) == 80
-
-    def test_description_too_long_raises(self) -> None:
-        """(d) description='x'*501 → max_length=500 違反 → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        with pytest.raises(ValidationError):
-            RoomCreate(name="X", workflow_id=uuid4(), description="x" * 501)
-
-    def test_prompt_kit_too_long_raises(self) -> None:
-        """(e) prompt_kit_prefix_markdown='x'*10001 → max_length=10000 違反 → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        with pytest.raises(ValidationError):
-            RoomCreate(name="X", workflow_id=uuid4(), prompt_kit_prefix_markdown="x" * 10001)
-
-    def test_extra_field_raises(self) -> None:
-        """(f) extra_field='z' → extra='forbid' → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        with pytest.raises(ValidationError):
-            RoomCreate.model_validate(
-                {"name": "X", "workflow_id": str(uuid4()), "extra_field": "z"}
-            )
-
-    def test_default_description_is_empty_string(self) -> None:
-        """description 省略 → デフォルト ''."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        schema = RoomCreate(name="X", workflow_id=uuid4())
-        assert schema.description == ""
-
-    def test_default_prompt_kit_is_empty_string(self) -> None:
-        """prompt_kit_prefix_markdown 省略 → デフォルト ''."""
-        from bakufu.interfaces.http.schemas.room import RoomCreate
-
-        schema = RoomCreate(name="X", workflow_id=uuid4())
-        assert schema.prompt_kit_prefix_markdown == ""
-
-
-# ---------------------------------------------------------------------------
-# TC-UT-RM-HTTP-002: RoomUpdate スキーマ検証
-# ---------------------------------------------------------------------------
-
-
-class TestRoomUpdateSchema:
-    """TC-UT-RM-HTTP-002: RoomUpdate — partial update with optional fields."""
-
-    def test_valid_name_passes(self) -> None:
-        """(a) name='新名前' → 通過."""
-        from bakufu.interfaces.http.schemas.room import RoomUpdate
-
-        schema = RoomUpdate(name="新名前")
-        assert schema.name == "新名前"
-
-    def test_name_none_passes(self) -> None:
-        """(b) name=None → 通過 (変更なし)."""
-        from bakufu.interfaces.http.schemas.room import RoomUpdate
-
-        schema = RoomUpdate(name=None)
-        assert schema.name is None
-
-    def test_empty_name_raises(self) -> None:
-        """(c) name='' → min_length=1 違反 → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import RoomUpdate
-
-        with pytest.raises(ValidationError):
-            RoomUpdate(name="")
-
-    def test_description_none_passes(self) -> None:
-        """(d) description=None → 通過 (変更なし)."""
-        from bakufu.interfaces.http.schemas.room import RoomUpdate
-
-        schema = RoomUpdate(description=None)
-        assert schema.description is None
-
-    def test_all_fields_none_passes(self) -> None:
-        """(e) 全フィールド None → 通過 (全変更なし)."""
-        from bakufu.interfaces.http.schemas.room import RoomUpdate
-
-        schema = RoomUpdate()
-        assert schema.name is None
-        assert schema.description is None
-        assert schema.prompt_kit_prefix_markdown is None
-
-    def test_extra_field_raises(self) -> None:
-        """(f) extra_field='z' → extra='forbid' → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import RoomUpdate
-
-        with pytest.raises(ValidationError):
-            RoomUpdate.model_validate({"name": "X", "extra_field": "z"})
-
-    def test_default_all_none(self) -> None:
-        """全フィールド省略時のデフォルトは None."""
-        from bakufu.interfaces.http.schemas.room import RoomUpdate
-
-        schema = RoomUpdate()
-        assert schema.name is None
-        assert schema.description is None
-        assert schema.prompt_kit_prefix_markdown is None
-
-
-# ---------------------------------------------------------------------------
-# TC-UT-RM-HTTP-003: AgentAssignRequest スキーマ検証
-# ---------------------------------------------------------------------------
-
-
-class TestAgentAssignRequestSchema:
-    """TC-UT-RM-HTTP-003: AgentAssignRequest role validation (_validate_role)."""
-
-    def test_valid_role_leader_passes(self) -> None:
-        """(a) agent_id=uuid4(), role='LEADER' → 通過."""
-        from bakufu.interfaces.http.schemas.room import AgentAssignRequest
-
-        schema = AgentAssignRequest(agent_id=uuid4(), role="LEADER")
-        assert schema.role == "LEADER"
-
-    def test_all_valid_roles_pass(self) -> None:
-        """_VALID_ROLES に含まれる全 10 値が通過することを確認。"""
-        from bakufu.interfaces.http.schemas.room import AgentAssignRequest, _VALID_ROLES  # type: ignore[attr-defined]
-
-        for role in _VALID_ROLES:
-            schema = AgentAssignRequest(agent_id=uuid4(), role=role)
-            assert schema.role == role
-
-    def test_empty_role_raises(self) -> None:
-        """(b) role='' → _validate_role: '' not in _VALID_ROLES → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import AgentAssignRequest
-
-        with pytest.raises(ValidationError):
-            AgentAssignRequest(agent_id=uuid4(), role="")
-
-    def test_role_too_long_raises(self) -> None:
-        """(c) role='x'*51 → max_length=50 or _validate_role → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import AgentAssignRequest
-
-        with pytest.raises(ValidationError):
-            AgentAssignRequest(agent_id=uuid4(), role="x" * 51)
-
-    def test_invalid_role_value_raises(self) -> None:
-        """Jensen 確認要件: assign_agent 無効 role → ValidationError (schema レベル, HTTP 422)."""
-        from bakufu.interfaces.http.schemas.room import AgentAssignRequest
-
-        with pytest.raises(ValidationError):
-            AgentAssignRequest(agent_id=uuid4(), role="NOT_A_ROLE")
-
-    def test_lowercase_role_raises(self) -> None:
-        """小文字 'leader' も _VALID_ROLES 外なので ValidationError."""
-        from bakufu.interfaces.http.schemas.room import AgentAssignRequest
-
-        with pytest.raises(ValidationError):
-            AgentAssignRequest(agent_id=uuid4(), role="leader")
-
-    def test_extra_field_raises(self) -> None:
-        """(d) extra_field='z' → extra='forbid' → ValidationError."""
-        from bakufu.interfaces.http.schemas.room import AgentAssignRequest
-
-        with pytest.raises(ValidationError):
-            AgentAssignRequest.model_validate(
-                {"agent_id": str(uuid4()), "role": "LEADER", "extra_field": "z"}
-            )
-
-
-# ---------------------------------------------------------------------------
-# TC-UT-RM-HTTP-004: RoomResponse / MemberResponse / RoomListResponse
-# ---------------------------------------------------------------------------
-
-
-class TestRoomResponseSchema:
-    """TC-UT-RM-HTTP-004: RoomResponse / MemberResponse / RoomListResponse シリアライズ."""
-
-    def _make_mock_membership(self, agent_id: object = None, role: str = "LEADER") -> Any:
-        """MemberResponse.model_validate に渡せる AgentMembership 風オブジェクトを返す."""
-        from unittest.mock import MagicMock
-
-        m = MagicMock()
-        m.agent_id = agent_id if agent_id is not None else uuid4()
-        m.role = role
-        m.joined_at = datetime.now(UTC)
-        return m
-
-    def _make_mock_room(
-        self,
-        name: str = "Vモデル開発室",
-        members: list[Any] | None = None,
-        archived: bool = False,
-    ) -> Any:
-        """RoomResponse._flatten_room が duck-typing で読み取る Room 風オブジェクト."""
-        from unittest.mock import MagicMock
-
-        r = MagicMock()
-        r.id = uuid4()
-        r.name = name
-        r.description = ""
-        r.workflow_id = uuid4()
-        r.members = members if members is not None else []
-        r.prompt_kit.prefix_markdown = "テストプロンプト"
-        r.archived = archived
-        return r
-
-    def test_room_response_id_is_str(self) -> None:
-        """model_validate(room) → id が str (UUID 文字列)."""
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        mock_room = self._make_mock_room()
-        resp = RoomResponse.model_validate(mock_room)
-        assert isinstance(resp.id, str)
-
-    def test_room_response_workflow_id_is_str(self) -> None:
-        """workflow_id が str (UUID 文字列)."""
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        mock_room = self._make_mock_room()
-        resp = RoomResponse.model_validate(mock_room)
-        assert isinstance(resp.workflow_id, str)
-
-    def test_room_response_archived_is_bool(self) -> None:
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        mock_room = self._make_mock_room(archived=True)
-        resp = RoomResponse.model_validate(mock_room)
-        assert resp.archived is True
-
-    def test_room_response_members_is_list(self) -> None:
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        mock_room = self._make_mock_room(members=[self._make_mock_membership()])
-        resp = RoomResponse.model_validate(mock_room)
-        assert isinstance(resp.members, list)
-        assert len(resp.members) == 1
-
-    def test_member_response_agent_id_is_str(self) -> None:
-        """MemberResponse.agent_id は str (UUID → str 変換)."""
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        agent_id = uuid4()
-        mock_room = self._make_mock_room(members=[self._make_mock_membership(agent_id=agent_id)])
-        resp = RoomResponse.model_validate(mock_room)
-        assert resp.members[0].agent_id == str(agent_id)
-
-    def test_member_response_role_is_str(self) -> None:
-        """MemberResponse.role は str."""
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        mock_room = self._make_mock_room(members=[self._make_mock_membership(role="REVIEWER")])
-        resp = RoomResponse.model_validate(mock_room)
-        assert resp.members[0].role == "REVIEWER"
-
-    def test_member_response_joined_at_is_iso8601_str(self) -> None:
-        """MemberResponse.joined_at は ISO 8601 str (_coerce_joined_at)."""
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        mock_room = self._make_mock_room(members=[self._make_mock_membership()])
-        resp = RoomResponse.model_validate(mock_room)
-        joined_at = resp.members[0].joined_at
-        assert isinstance(joined_at, str)
-        assert "T" in joined_at  # ISO 8601 contains "T" separator
-
-    def test_room_response_prompt_kit_prefix_markdown_flattened(self) -> None:
-        """prompt_kit.prefix_markdown が prompt_kit_prefix_markdown にフラット化される (_flatten_room)."""
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        mock_room = self._make_mock_room()
-        resp = RoomResponse.model_validate(mock_room)
-        assert resp.prompt_kit_prefix_markdown == "テストプロンプト"
-
-    def test_room_list_response_total_matches_items_len(self) -> None:
-        """RoomListResponse.total が len(items) と一致。"""
-        from bakufu.interfaces.http.schemas.room import RoomListResponse, RoomResponse
-
-        rooms = [RoomResponse.model_validate(self._make_mock_room(name=f"部屋{i}")) for i in range(3)]
-        resp = RoomListResponse(items=rooms, total=len(rooms))
-        assert resp.total == 3
-        assert len(resp.items) == 3
-
-    def test_room_response_extra_field_raises(self) -> None:
-        """extra='forbid': 未知フィールドは ValidationError."""
-        from bakufu.interfaces.http.schemas.room import RoomResponse
-
-        with pytest.raises(ValidationError):
-            RoomResponse.model_validate(
-                {
-                    "id": str(uuid4()),
-                    "name": "X",
-                    "description": "",
-                    "workflow_id": str(uuid4()),
-                    "members": [],
-                    "prompt_kit_prefix_markdown": "",
-                    "archived": False,
-                    "extra": "z",
-                }
-            )
-
-
-# ---------------------------------------------------------------------------
-# TC-UT-RM-HTTP-005: room_invariant_violation_handler [FAIL]/Next: 除去 + kind 分岐
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-class TestRoomInvariantViolationHandler:
-    """TC-UT-RM-HTTP-005: room_invariant_violation_handler 前処理ルール + kind 分岐 (確定 C)."""
-
-    def _make_request(self) -> Any:
-        return MagicMock()
-
-    async def test_name_range_returns_422(self) -> None:
-        """(a) kind='name_range' → HTTP 422."""
-        from bakufu.domain.exceptions import RoomInvariantViolation
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        exc = RoomInvariantViolation(
-            kind="name_range",
-            message=(
-                "[FAIL] Room name は 1〜80 文字でなければなりません。"
-                "\nNext: 1〜80 文字の名前を指定してください。"
-            ),
-        )
-        resp = await room_invariant_violation_handler(self._make_request(), exc)  # type: ignore[arg-type]
-        assert resp.status_code == 422  # type: ignore[union-attr]
-
-    async def test_name_range_error_code_is_validation_error(self) -> None:
-        import json
-
-        from bakufu.domain.exceptions import RoomInvariantViolation
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        exc = RoomInvariantViolation(
-            kind="name_range",
-            message=(
-                "[FAIL] Room name は 1〜80 文字でなければなりません。"
-                "\nNext: 1〜80 文字の名前を指定してください。"
-            ),
-        )
-        resp = await room_invariant_violation_handler(self._make_request(), exc)  # type: ignore[arg-type]
-        body = json.loads(resp.body)  # type: ignore[union-attr]
-        assert body["error"]["code"] == "validation_error"
-
-    async def test_name_range_removes_fail_prefix(self) -> None:
-        """[FAIL] プレフィックスが除去され message に含まれないこと (確定 C)."""
-        import json
-
-        from bakufu.domain.exceptions import RoomInvariantViolation
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        exc = RoomInvariantViolation(
-            kind="name_range",
-            message=(
-                "[FAIL] Room name は 1〜80 文字でなければなりません。"
-                "\nNext: 1〜80 文字の名前を指定してください。"
-            ),
-        )
-        resp = await room_invariant_violation_handler(self._make_request(), exc)  # type: ignore[arg-type]
-        body = json.loads(resp.body)  # type: ignore[union-attr]
-        assert "[FAIL]" not in body["error"]["message"]
-
-    async def test_name_range_removes_next_suffix(self) -> None:
-        """\\nNext:.* サフィックスが除去されること (確定 C)."""
-        import json
-
-        from bakufu.domain.exceptions import RoomInvariantViolation
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        exc = RoomInvariantViolation(
-            kind="name_range",
-            message=(
-                "[FAIL] Room name は 1〜80 文字でなければなりません。"
-                "\nNext: 1〜80 文字の名前を指定してください。"
-            ),
-        )
-        resp = await room_invariant_violation_handler(self._make_request(), exc)  # type: ignore[arg-type]
-        body = json.loads(resp.body)  # type: ignore[union-attr]
-        assert "Next:" not in body["error"]["message"]
-
-    async def test_name_range_message_is_clean_business_text(self) -> None:
-        """期待 message: 'Room name は 1〜80 文字でなければなりません。' (前処理後の本文のみ)."""
-        import json
-
-        from bakufu.domain.exceptions import RoomInvariantViolation
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        exc = RoomInvariantViolation(
-            kind="name_range",
-            message=(
-                "[FAIL] Room name は 1〜80 文字でなければなりません。"
-                "\nNext: 1〜80 文字の名前を指定してください。"
-            ),
-        )
-        resp = await room_invariant_violation_handler(self._make_request(), exc)  # type: ignore[arg-type]
-        body = json.loads(resp.body)  # type: ignore[union-attr]
-        assert body["error"]["message"] == "Room name は 1〜80 文字でなければなりません。"
-
-    async def test_member_not_found_returns_404(self) -> None:
-        """(b) kind='member_not_found' → HTTP 404 (MSG-RM-HTTP-005 分岐確認)."""
-        from bakufu.domain.exceptions import RoomInvariantViolation
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        exc = RoomInvariantViolation(
-            kind="member_not_found",
-            message="[FAIL] Member not found: agent_id=..., role=LEADER\nNext: ...",
-        )
-        resp = await room_invariant_violation_handler(self._make_request(), exc)  # type: ignore[arg-type]
-        assert resp.status_code == 404  # type: ignore[union-attr]
-
-    async def test_member_not_found_error_code(self) -> None:
-        """kind='member_not_found' → error.code='not_found'."""
-        import json
-
-        from bakufu.domain.exceptions import RoomInvariantViolation
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        exc = RoomInvariantViolation(
-            kind="member_not_found",
-            message="[FAIL] Member not found\nNext: ...",
-        )
-        resp = await room_invariant_violation_handler(self._make_request(), exc)  # type: ignore[arg-type]
-        body = json.loads(resp.body)  # type: ignore[union-attr]
-        assert body["error"]["code"] == "not_found"
-
-    async def test_member_not_found_error_message(self) -> None:
-        """kind='member_not_found' → MSG-RM-HTTP-005 固定文言 (確定 C)."""
-        import json
-
-        from bakufu.domain.exceptions import RoomInvariantViolation
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        exc = RoomInvariantViolation(
-            kind="member_not_found",
-            message="[FAIL] Member not found\nNext: ...",
-        )
-        resp = await room_invariant_violation_handler(self._make_request(), exc)  # type: ignore[arg-type]
-        body = json.loads(resp.body)  # type: ignore[union-attr]
-        assert body["error"]["message"] == "Agent membership not found in this room."
-
-    async def test_wrong_type_raises_type_error(self) -> None:
-        """非 RoomInvariantViolation → TypeError (Fail Fast 確認)."""
-        from bakufu.interfaces.http.error_handlers import room_invariant_violation_handler
-
-        with pytest.raises(TypeError, match="Expected RoomInvariantViolation"):
-            await room_invariant_violation_handler(self._make_request(), ValueError("oops"))  # type: ignore[arg-type]
-
-
-# ---------------------------------------------------------------------------
-# TC-UT-RM-HTTP-006: RoomService.__init__ 4 repo 構造
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-class TestRoomServiceInit:
-    """TC-UT-RM-HTTP-006: RoomService.__init__ 4 repo 構造 (detailed-design.md §確定G)."""
-
-    def _make_mock_session(self) -> Any:
-        """begin() が async context-manager を返す MagicMock session."""
-        from unittest.mock import AsyncMock
-
-        mock_cm = MagicMock()
-        mock_cm.__aenter__ = AsyncMock(return_value=None)
-        mock_cm.__aexit__ = AsyncMock(return_value=False)
-        mock_session = MagicMock()
-        mock_session.begin = MagicMock(return_value=mock_cm)
-        return mock_session
-
-    async def test_service_constructs_successfully(self) -> None:
-        from bakufu.application.services.room_service import RoomService
-
-        service = RoomService(
-            room_repo=MagicMock(),
-            empire_repo=MagicMock(),
-            workflow_repo=MagicMock(),
-            agent_repo=MagicMock(),
-            session=self._make_mock_session(),
-        )
-        assert service is not None
-
-    async def test_service_stores_room_repo(self) -> None:
-        """_room_repo に room_repo が格納される."""
-        from bakufu.application.services.room_service import RoomService
-
-        mock_room_repo = MagicMock()
-        service = RoomService(
-            room_repo=mock_room_repo,
-            empire_repo=MagicMock(),
-            workflow_repo=MagicMock(),
-            agent_repo=MagicMock(),
-            session=self._make_mock_session(),
-        )
-        assert service._room_repo is mock_room_repo  # pyright: ignore[reportPrivateUsage]
-
-    async def test_service_stores_empire_repo(self) -> None:
-        """_empire_repo に empire_repo が格納される."""
-        from bakufu.application.services.room_service import RoomService
-
-        mock_empire_repo = MagicMock()
-        service = RoomService(
-            room_repo=MagicMock(),
-            empire_repo=mock_empire_repo,
-            workflow_repo=MagicMock(),
-            agent_repo=MagicMock(),
-            session=self._make_mock_session(),
-        )
-        assert service._empire_repo is mock_empire_repo  # pyright: ignore[reportPrivateUsage]
-
-    async def test_service_stores_workflow_repo(self) -> None:
-        """_workflow_repo に workflow_repo が格納される."""
-        from bakufu.application.services.room_service import RoomService
-
-        mock_workflow_repo = MagicMock()
-        service = RoomService(
-            room_repo=MagicMock(),
-            empire_repo=MagicMock(),
-            workflow_repo=mock_workflow_repo,
-            agent_repo=MagicMock(),
-            session=self._make_mock_session(),
-        )
-        assert service._workflow_repo is mock_workflow_repo  # pyright: ignore[reportPrivateUsage]
-
-    async def test_service_stores_agent_repo(self) -> None:
-        """_agent_repo に agent_repo が格納される."""
-        from bakufu.application.services.room_service import RoomService
-
-        mock_agent_repo = MagicMock()
-        service = RoomService(
-            room_repo=MagicMock(),
-            empire_repo=MagicMock(),
-            workflow_repo=MagicMock(),
-            agent_repo=mock_agent_repo,
-            session=self._make_mock_session(),
-        )
-        assert service._agent_repo is mock_agent_repo  # pyright: ignore[reportPrivateUsage]
-
-
-# ---------------------------------------------------------------------------
-# TC-UT-RM-HTTP-010: interfaces/http/ 依存方向静的解析 — ast.walk() 全ノード走査
-# ---------------------------------------------------------------------------
-
-
-class TestStaticDependencyAnalysisRoom:
-    """TC-UT-RM-HTTP-010: routers/ と schemas/ の依存方向を ast.walk() で全検査.
-
-    TC-UT-EM-HTTP-010 (empire) の同一パターン。
-    旧実装 (tree.body のみ走査) は関数内遅延 import を見逃す盲点があった。
-    ``ast.walk(ast.parse(src))`` で全 AST ノード (関数内・クラス内・ネスト含む) を
-    走査し、``bakufu.domain`` / ``bakufu.infrastructure`` への import が
-    存在しないことを確認する。
-
-    スコープ: interfaces/http/routers/ + interfaces/http/schemas/
-    除外: app.py / dependencies.py / error_handlers.py (設計上 domain import が正当)
-    """
-
-    def _interfaces_http_dir(self) -> Path:
-        import bakufu.interfaces.http.app as _app_mod
-
-        return Path(_app_mod.__file__).parent  # type: ignore[arg-type]
-
-    def _collect_all_imports(self, py_file: Path) -> list[tuple[str, int]]:
-        """ast.walk() でファイル内の全 import 文 (関数内遅延 import 含む) を収集する.
-
-        旧 ``_collect_toplevel_imports`` は ``tree.body`` のみを走査していたが、
-        本メソッドは ``ast.walk(tree)`` で AST 全ノードを走査するため、
-        関数本体・クラス本体・ネストされたスコープ内の import も捕捉できる。
-        """
-        tree = ast.parse(py_file.read_text(encoding="utf-8"))
-        results: list[tuple[str, int]] = []
-        for node in ast.walk(tree):
-            if isinstance(node, ast.ImportFrom):
-                module = node.module or ""
-                results.append((module, node.lineno))
-            elif isinstance(node, ast.Import):
-                for alias in node.names:
-                    results.append((alias.name, node.lineno))
-        return results
-
-    def test_room_router_has_no_domain_import(self) -> None:
-        """routers/ と schemas/ は bakufu.domain を一切 import してはならない (遅延含む).
-
-        ast.walk() で関数内遅延 import も検出する。
-        """
-        interfaces_dir = self._interfaces_http_dir()
-        violations: list[str] = []
-        scan_dirs = [interfaces_dir / "routers", interfaces_dir / "schemas"]
-        for scan_dir in scan_dirs:
-            for py_file in sorted(scan_dir.rglob("*.py")):
-                for module_name, lineno in self._collect_all_imports(py_file):
-                    if module_name.startswith("bakufu.domain"):
-                        violations.append(f"{py_file.name}:{lineno}: import of {module_name}")
-        assert violations == [], (
-            "bakufu.domain imports detected in routers/ or schemas/ (including deferred):\n"
-            + "\n".join(violations)
-        )
-
-    def test_room_router_has_no_infrastructure_import(self) -> None:
-        """routers/ と schemas/ は bakufu.infrastructure を一切 import してはならない (遅延含む).
-
-        ast.walk() で関数内遅延 import も検出する。
-        """
-        interfaces_dir = self._interfaces_http_dir()
-        violations: list[str] = []
-        scan_dirs = [interfaces_dir / "routers", interfaces_dir / "schemas"]
-        for scan_dir in scan_dirs:
-            for py_file in sorted(scan_dir.rglob("*.py")):
-                for module_name, lineno in self._collect_all_imports(py_file):
-                    if module_name.startswith("bakufu.infrastructure"):
-                        violations.append(f"{py_file.name}:{lineno}: import of {module_name}")
-        assert violations == [], (
-            "bakufu.infrastructure imports detected in routers/ or schemas/ (including deferred):\n"
-            + "\n".join(violations)
-        )
